Removing flight and alert/status naming
