+++ src/Components/ADT3DSceneBuilder/Internal/Behaviors/BehaviorsForm.tsx
@@ -49,8 +49,6 @@ import {
     ITwinToObjectMapping
 } from '../../../../Models/Types/Generated/3DScenesConfiguration-v1.0.0';
 import ViewerConfigUtility from '../../../../Models/Classes/ViewerConfigUtility';
-import AlertsTab from './Internal/AlertsTab';
-import StatusTab from './Internal/StatusTab';
 import WidgetsTab from './Internal/WidgetsTab';
 import {
     BehaviorFormReducer,
@@ -69,23 +67,16 @@ import {
     BehaviorFormContextProvider
 } from '../../../../Models/Context/BehaviorFormContext/BehaviorFormContext';
 import { BehaviorFormContextActionType } from '../../../../Models/Context/BehaviorFormContext/BehaviorFormContext.types';
-import { LOCAL_STORAGE_KEYS } from '../../../../Models/Constants';
 import { VisualRulesTab } from './Internal/VisualRulesTab';
 import VisualRuleForm from '../VisualRuleForm/VisualRuleForm';
 
-const showVisualRulesFeature =
-    localStorage.getItem(
-        LOCAL_STORAGE_KEYS.FeatureFlags.VisualRules.showVisualRulesFeature
-    ) === 'true' || false;
 const getElementsFromBehavior = (behavior: IBehavior) =>
     behavior.datasources.filter(
         ViewerConfigUtility.isElementTwinToObjectMappingDataSource
     )[0] || null;
 
 enum BehaviorPivot {
-    alerts = 'alerts',
     elements = 'elements',
-    states = 'states',
     twins = 'twins',
     visualRules = 'visualRules',
     widgets = 'widgets'
@@ -603,56 +594,16 @@ const SceneBehaviorsForm: React.FC<IADT3DSceneBuilderBehaviorFormProps> = ({
                                 selectedElements={selectedElements}
                             />
                         </PivotItem>
-                        {!showVisualRulesFeature && (
-                            <PivotItem
-                                className={commonPanelStyles.formTabContents}
-                                headerText={t('3dSceneBuilder.statesTab')}
-                                itemKey={BehaviorPivot.states}
-                                onRenderItemLink={(props, defaultRenderer) =>
-                                    setPivotToRequired(
-                                        behaviorState.validityMap?.get('Status')
-                                            ?.isValid,
-                                        t,
-                                        props,
-                                        defaultRenderer
-                                    )
-                                }
-                            >
-                                <StatusTab
-                                    onValidityChange={onTabValidityChange}
-                                />
-                            </PivotItem>
-                        )}
-                        {!showVisualRulesFeature && (
-                            <PivotItem
-                                className={commonPanelStyles.formTabContents}
-                                headerText={t('3dSceneBuilder.alertsTab')}
-                                itemKey={BehaviorPivot.alerts}
-                                onRenderItemLink={(props, defaultRenderer) =>
-                                    setPivotToRequired(
-                                        behaviorState.validityMap?.get('Alerts')
-                                            ?.isValid,
-                                        t,
-                                        props,
-                                        defaultRenderer
-                                    )
-                                }
-                            >
-                                <AlertsTab />
-                            </PivotItem>
-                        )}
-                        {showVisualRulesFeature && (
-                            <PivotItem
-                                className={commonPanelStyles.formTabContents}
-                                headerText={t('3dSceneBuilder.visualRulesTab')}
-                                itemKey={BehaviorPivot.visualRules}
-                            >
-                                <VisualRulesTab
-                                    onEditRule={onEditRule}
-                                    onAddRule={onAddRule}
-                                />
-                            </PivotItem>
-                        )}
+                        <PivotItem
+                            className={commonPanelStyles.formTabContents}
+                            headerText={t('3dSceneBuilder.visualRulesTab')}
+                            itemKey={BehaviorPivot.visualRules}
+                        >
+                            <VisualRulesTab
+                                onEditRule={onEditRule}
+                                onAddRule={onAddRule}
+                            />
+                        </PivotItem>
                         <PivotItem
                             className={commonPanelStyles.formTabContents}
                             headerText={t('3dSceneBuilder.widgets')}
