+++ src/Components/BehaviorsModal/Internal/BehaviorSection/BehaviorSection.tsx
@@ -1,48 +1,95 @@
-import { Icon, Stack } from '@fluentui/react';
-import React, { useContext, useMemo } from 'react';
+import { Callout, DirectionalHint, Stack } from '@fluentui/react';
+import React, { useCallback, useContext, useMemo, useState } from 'react';
 import ViewerConfigUtility from '../../../../Models/Classes/ViewerConfigUtility';
-import { BehaviorModalMode } from '../../../../Models/Constants';
 import {
-    wrapTextInTemplateString,
-    parseLinkedTwinExpression,
-    stripTemplateStringsFromText
-} from '../../../../Models/Services/Utils';
-import {
-    IBehavior,
-    IExpressionRangeVisual
-} from '../../../../Models/Types/Generated/3DScenesConfiguration-v1.0.0';
-import { getElementsPanelAlertStyles } from '../../../ElementsPanel/ViewerElementsPanel.styles';
-import { StatusPills } from '../../../StatusPills/StatusPills';
+    shouldShowVisual,
+    hasBadge
+} from '../../../../Models/SharedUtils/VisualRuleUtils';
+import ColorPills from '../../../ColorPills/ColorPills';
 import { BehaviorsModalContext } from '../../BehaviorsModal';
 import WidgetsContainer from '../Widgets/WidgetsContainer';
-import { getStyles } from './BehaviorSection.styles';
-
-export interface IBehaviorsSectionProps {
-    behavior: IBehavior;
-}
+import { getCalloutStyles, getStyles } from './BehaviorSection.styles';
+import { useId } from '@fluentui/react-hooks';
+import ColorPillsCalloutContent from '../../../ColorPillsCalloutContent/ColorPillsCalloutContent';
+import { IBehaviorsSectionProps } from './BehaviorSection.types';
+import BadgeBlock from './BadgeBlock';
+import {
+    VisualBadges,
+    VisualColorings
+} from '../../../../Models/Constants/VisualRuleTypes';
+import { useExtendedTheme } from '../../../../Models/Hooks/useExtendedTheme';
+import { BehaviorModalMode } from '../../../../Models/Constants/Enums';
 
 const BehaviorSection: React.FC<IBehaviorsSectionProps> = ({ behavior }) => {
     const styles = getStyles();
     const { twins, mode } = useContext(BehaviorsModalContext);
+    const badgeIdPrefix = useId('cb-visual-rule-badge');
+    const stackId = useId('cb-behavior-section-header-tooltip');
+    const theme = useExtendedTheme();
 
-    const alertVisuals = useMemo(() => {
-        let visibleAlertVisuals =
-            behavior.visuals.filter(ViewerConfigUtility.isAlertVisual) || [];
+    // State
+    const [isCalloutOpen, setIsCalloutOpen] = useState(false);
 
-        if (mode !== BehaviorModalMode.preview) {
-            visibleAlertVisuals = visibleAlertVisuals.filter((av) =>
-                parseLinkedTwinExpression(av.valueExpression, twins)
-            );
-        }
-        return visibleAlertVisuals;
+    const onItemHover = useCallback(() => {
+        setIsCalloutOpen(true);
+    }, []);
+
+    const onItemBlur = useCallback(() => {
+        setIsCalloutOpen(false);
+    }, []);
+
+    const { meshColorings, badges } = useMemo(() => {
+        const rules =
+            behavior.visuals.filter(ViewerConfigUtility.isVisualRule) || [];
+        const visualColorings: VisualColorings[] = [];
+        const visualBadges: VisualBadges[] = [];
+        rules.forEach((rule) => {
+            rule.valueRanges.forEach((condition) => {
+                if (mode === BehaviorModalMode.preview) {
+                    // Remove filtering if in preview mode
+                    if (hasBadge(condition)) {
+                        visualBadges.push({
+                            color: condition.visual.color,
+                            iconName: condition.visual.iconName,
+                            labelExpression: condition.visual.labelExpression
+                        });
+                    } else {
+                        visualColorings.push({
+                            color: condition.visual.color,
+                            label: condition.visual.labelExpression
+                        });
+                    }
+                } else {
+                    if (
+                        shouldShowVisual(
+                            rule.valueRangeType,
+                            twins,
+                            rule.valueExpression,
+                            condition.values
+                        )
+                    ) {
+                        if (hasBadge(condition)) {
+                            visualBadges.push({
+                                color: condition.visual.color,
+                                iconName: condition.visual.iconName,
+                                labelExpression:
+                                    condition.visual.labelExpression
+                            });
+                        } else {
+                            visualColorings.push({
+                                color: condition.visual.color,
+                                label: condition.visual.labelExpression
+                            });
+                        }
+                    }
+                }
+            });
+        });
+        return {
+            meshColorings: visualColorings,
+            badges: visualBadges
+        };
     }, [behavior.visuals, mode, twins]);
-
-    const statusVisuals = useMemo(
-        () =>
-            behavior.visuals.filter(ViewerConfigUtility.isStatusColorVisual) ||
-            [],
-        [behavior]
-    );
 
     const popoverVisual = useMemo(
         () => behavior.visuals.filter(ViewerConfigUtility.isPopoverVisual)[0],
@@ -50,59 +97,50 @@ const BehaviorSection: React.FC<IBehaviorsSectionProps> = ({ behavior }) => {
     );
 
     return (
-        <div className={styles.behaviorSection}>
-            <Stack
-                horizontal={true}
-                verticalAlign={'center'}
-                disableShrink={true}
-                className={styles.behaviorHeader}
-            >
-                <StatusPills
-                    statusVisuals={statusVisuals}
-                    width={'compact'}
-                    twins={twins}
-                />
-                {behavior.displayName}
-            </Stack>
-            {alertVisuals.map((av, idx) => (
-                <AlertBlock alertVisual={av} key={`${av.type}-${idx}`} />
-            ))}
-            {popoverVisual && (
-                <WidgetsContainer popoverVisual={popoverVisual} />
-            )}
-        </div>
-    );
-};
-
-const AlertBlock: React.FC<{ alertVisual: IExpressionRangeVisual }> = ({
-    alertVisual
-}) => {
-    if (!alertVisual && !alertVisual.valueRanges[0]) {
-        return null;
-    }
-
-    const styles = getStyles();
-    const {
-        visual: { color, iconName, labelExpression }
-    } = alertVisual.valueRanges[0];
-
-    const alertStyles = getElementsPanelAlertStyles(color);
-    const { twins, mode } = useContext(BehaviorsModalContext);
-
-    return (
-        <div className={styles.infoContainer}>
-            <div className={alertStyles.alertCircle}>
-                <Icon iconName={iconName} />
+        <>
+            <div className={styles.behaviorSection}>
+                <Stack
+                    id={stackId}
+                    horizontal={true}
+                    verticalAlign={'center'}
+                    disableShrink={true}
+                    className={styles.behaviorHeader}
+                    onFocus={onItemHover}
+                    onMouseEnter={onItemHover}
+                    onBlur={onItemBlur}
+                    onMouseLeave={onItemBlur}
+                    tabIndex={0}
+                >
+                    <ColorPills
+                        visualColorings={meshColorings}
+                        width={'compact'}
+                    />
+                    {behavior.displayName}
+                    {isCalloutOpen && (
+                        <Callout
+                            target={`#${stackId}`}
+                            directionalHint={DirectionalHint.rightTopEdge}
+                            gapSpace={24}
+                            isBeakVisible={false}
+                            styles={getCalloutStyles(theme)}
+                        >
+                            <ColorPillsCalloutContent
+                                visualColorings={meshColorings}
+                            />
+                        </Callout>
+                    )}
+                </Stack>
+                {badges.map((bv, idx) => (
+                    <BadgeBlock
+                        badgeVisual={bv}
+                        key={`${badgeIdPrefix}-${idx}`}
+                    />
+                ))}
+                {popoverVisual && (
+                    <WidgetsContainer popoverVisual={popoverVisual} />
+                )}
             </div>
-            <div className={styles.infoTextContainer}>
-                {mode === BehaviorModalMode.preview
-                    ? stripTemplateStringsFromText(labelExpression)
-                    : parseLinkedTwinExpression(
-                          wrapTextInTemplateString(labelExpression),
-                          twins
-                      )}
-            </div>
-        </div>
+        </>
     );
 };
