+++ src/Components/ElementsPanel/ViewerElementsPanel.tsx
@@ -1,5 +1,4 @@
 import React, { memo, useMemo, useRef, useState } from 'react';
-import ElementList from './Internal/ElementsList';
 import { useTranslation } from 'react-i18next';
 import { Icon, SearchBox, Spinner, SpinnerSize } from '@fluentui/react';
 import { IViewerElementsPanelProps } from './ViewerElementsPanel.types';
@@ -10,13 +9,12 @@ import {
     wrapTextInTemplateString,
     parseLinkedTwinExpression
 } from '../../Models/Services/Utils';
-import VisualRuleElementsList from './Internal/VisualRuleElementsList';
-import { LOCAL_STORAGE_KEYS } from '../../Models/Constants/Constants';
+import ElementsList from './Internal/ElementsList';
 
-const showVisualRulesFeature =
-    localStorage.getItem(
-        LOCAL_STORAGE_KEYS.FeatureFlags.VisualRules.showVisualRulesFeature
-    ) === 'true';
+// const showVisualRulesFeature =
+//     localStorage.getItem(
+//         LOCAL_STORAGE_KEYS.FeatureFlags.VisualRules.showVisualRulesFeature
+//     ) === 'true';
 
 const ViewerElementsPanel: React.FC<IViewerElementsPanelProps> = ({
     panelItems,
@@ -94,26 +92,15 @@ const ViewerElementsPanel: React.FC<IViewerElementsPanelProps> = ({
                         value={filterTerm}
                         className={elementsPanelStyles.filterBox}
                     />
-                    {showVisualRulesFeature ? (
-                        <VisualRuleElementsList
-                            isLoading={isLoading}
-                            isModal={false}
-                            panelItems={filteredPanelItems}
-                            onItemClick={onItemClick}
-                            onItemHover={onItemHover}
-                            onItemBlur={onItemBlur}
-                            filterTerm={filterTerm}
-                        />
-                    ) : (
-                        <ElementList
-                            isLoading={isLoading}
-                            panelItems={filteredPanelItems}
-                            onItemClick={onItemClick}
-                            onItemHover={onItemHover}
-                            onItemBlur={onItemBlur}
-                            filterTerm={filterTerm}
-                        />
-                    )}
+                    <ElementsList
+                        isLoading={isLoading}
+                        isModal={false}
+                        panelItems={filteredPanelItems}
+                        onItemClick={onItemClick}
+                        onItemHover={onItemHover}
+                        onItemBlur={onItemBlur}
+                        filterTerm={filterTerm}
+                    />
                 </div>
             </Draggable>
         </div>
