+++ src/Components/VisualsModal/VisualsModal.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
-import { LOCAL_STORAGE_KEYS } from '../../Models/Constants/Constants';
 import {
     IBehavior,
     ITwinToObjectMapping,
     IVisual
 } from '../../Models/Types/Generated/3DScenesConfiguration-v1.0.0';
 import ElementsList from '../ElementsPanel/Internal/ElementsList';
-import VisualRuleElementsList from '../ElementsPanel/Internal/VisualRuleElementsList';
 import { IViewerElementsPanelItem } from '../ElementsPanel/ViewerElementsPanel.types';
-import { getStyles } from './AlertModal.styles';
+import { getStyles } from './VisualsModal.styles';
 
-export interface IAlertModalProps {
+export interface IVisualsModalProps {
     alerts: IViewerElementsPanelItem;
     onClose: () => any;
     position: { left: number; top: number };
@@ -31,12 +29,7 @@ export interface IAlertModalProps {
     ) => void;
 }
 
-const showVisualRulesFeature =
-    localStorage.getItem(
-        LOCAL_STORAGE_KEYS.FeatureFlags.VisualRules.showVisualRulesFeature
-    ) === 'true';
-
-const AlertModal: React.FC<IAlertModalProps> = ({
+const VisualsModal: React.FC<IVisualsModalProps> = ({
     onClose,
     position,
     alerts,
@@ -52,26 +45,16 @@ const AlertModal: React.FC<IAlertModalProps> = ({
             className={styles.boundaryLayer + ' cb-base-fade-in'}
             onMouseLeave={onClose}
         >
-            {showVisualRulesFeature ? (
-                <VisualRuleElementsList
-                    isLoading={false}
-                    isModal={true}
-                    panelItems={alerts?.element ? [alerts] : []}
-                    onItemClick={onItemClick}
-                    onItemBlur={onItemBlur}
-                    onItemHover={onItemHover}
-                />
-            ) : (
-                <ElementsList
-                    isLoading={false}
-                    panelItems={alerts?.element ? [alerts] : []}
-                    onItemClick={onItemClick}
-                    onItemBlur={onItemBlur}
-                    onItemHover={onItemHover}
-                />
-            )}
+            <ElementsList
+                isLoading={false}
+                isModal={true}
+                panelItems={alerts?.element ? [alerts] : []}
+                onItemClick={onItemClick}
+                onItemBlur={onItemBlur}
+                onItemHover={onItemHover}
+            />
         </div>
     );
 };
 
-export default React.memo(AlertModal);
+export default React.memo(VisualsModal);
