+++ src/Models/Context/BehaviorFormContext/BehaviorFormContext.types.ts
@@ -2,7 +2,6 @@ import {
     IBehavior,
     IElementTwinToObjectMappingDataSource,
     IExpressionRangeVisual,
-    IValueRange,
     IWidget
 } from '../../Types/Generated/3DScenesConfiguration-v1.0.0';
 
@@ -36,15 +35,6 @@ export enum BehaviorFormContextActionType {
     FORM_BEHAVIOR_VISUAL_RULE_ADD_OR_UPDATE = 'FORM_BEHAVIOR_VISUAL_RULE_ADD_OR_UPDATE',
     FORM_BEHAVIOR_VISUAL_RULE_REMOVE = 'FORM_BEHAVIOR_VISUAL_RULE_REMOVE',
 
-    /** add or update an alert visual
-     * @deprecated
-     */
-    FORM_BEHAVIOR_ALERT_VISUAL_ADD_OR_UPDATE = 'FORM_BEHAVIOR_ALERT_VISUAL_ADD_OR_UPDATE',
-    /** remove the alert visual
-     * @deprecated
-     */
-    FORM_BEHAVIOR_ALERT_VISUAL_REMOVE = 'FORM_BEHAVIOR_ALERT_VISUAL_REMOVE',
-
     /** add or update a twin alias */
     FORM_BEHAVIOR_ALIAS_ADD = 'FORM_BEHAVIOR_ALIAS_ADD',
     /** remove a twin alias */
@@ -64,18 +54,6 @@ export enum BehaviorFormContextActionType {
 
     /** reverts all changes to the behavior back to it's initial state */
     FORM_BEHAVIOR_RESET = 'FORM_BEHAVIOR_RESET',
-    /** Add or update a status visual
-     * @deprecated
-     */
-    FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE = 'FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE',
-    /** Update the value ranges for the status visual
-     * @deprecated
-     */
-    FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE_RANGES = 'FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE_RANGES',
-    /** remove a status visual
-     * @deprecated
-     */
-    FORM_BEHAVIOR_STATUS_VISUAL_REMOVE = 'FORM_BEHAVIOR_STATUS_VISUAL_REMOVE',
     /** add or update a widget */
     FORM_BEHAVIOR_WIDGET_ADD_OR_UPDATE = 'FORM_BEHAVIOR_WIDGET_ADD_OR_UPDATE',
     /** remove a widget */
@@ -84,14 +62,6 @@ export enum BehaviorFormContextActionType {
 
 /** The actions to update the state */
 export type BehaviorFormContextAction =
-    // ALERTS - DEPRECATE IN FAVOR OF VISUAL RULES
-    | {
-          type: BehaviorFormContextActionType.FORM_BEHAVIOR_ALERT_VISUAL_ADD_OR_UPDATE;
-          payload: { visual: IExpressionRangeVisual };
-      }
-    | {
-          type: BehaviorFormContextActionType.FORM_BEHAVIOR_ALERT_VISUAL_REMOVE;
-      }
     // ALIASES
     | {
           type: BehaviorFormContextActionType.FORM_BEHAVIOR_ALIAS_ADD;
@@ -134,20 +104,6 @@ export type BehaviorFormContextAction =
               behavior?: IBehavior;
               layerIds?: string[];
           };
-      }
-    // STATUS - DEPRECATE IN FAVOR OF VISUAL RULES
-    | {
-          type: BehaviorFormContextActionType.FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE;
-          payload: { visual: IExpressionRangeVisual };
-      }
-    | {
-          type: BehaviorFormContextActionType.FORM_BEHAVIOR_STATUS_VISUAL_ADD_OR_UPDATE_RANGES;
-          payload: {
-              ranges: IValueRange[];
-          };
-      }
-    | {
-          type: BehaviorFormContextActionType.FORM_BEHAVIOR_STATUS_VISUAL_REMOVE;
       }
     // WIDGETS
     | {
